[avalanche] Don't register known dangling proofs
Summary:
After a dangling proof has been evicted, the node should not accept it again unless it finally has a node pending for it. This prevents from downloading again a proof that never gets attached to a node.

The proof ids are stored in a rolling bloom filter so the memory usage is bound. Since a false positive in that filter is not critical, the parameters are chosen to keep the memory usage low (about 10kB).

Test Plan:
  ninja all check-extended

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11738