[Cashtab] Add parseInvalidSettingsForMigration function
Summary: Related to T2506. Function looks at the settings in the localStorage obj, determines if there is a missing parameter, then adds the default setting for only that parameter before restesting isValidCashtabSettings.

Test Plan:
Ensure that at least one setting is not default
in Ticker.js, create a testSetting and set its value to be either true or false
On save, refresh and check the application's memory in devtools
Existing settings should maintain their user defined values, while testSetting should be set to the default value

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11740