Use designated initializers for ChainstateManager::Options and move  into kernel:: namespace
Summary:
> Move ChainstateManagerOpts into kernel:: namespace
>
> It should have been there in the first place.

> Use designated initializers for ChainstateManager::Options
>
> This wasn't available at the time when ChainstateManager::Options was
> introduced but is helpful to be explicit and ensure correctness.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/25607 | core#25607]]

Note that designated initializers are a c++20 feature, but are supported out of the box with C++17 by all supported gcc and clang versions.
For Visual Studio 2022 it is supported with `/std:c++20` language mode.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15312