[avalanche] Use proof dust threshold in unit tests instead of low coin amounts
Summary:
This makes it much easier to bump PROOF_DUST_THRESHOLD without breaking tons of tests.
I did not update every conceivable Amount, but focused on those that will break if the
dust threshold is bumped by a factor of 10 to 100. Any further tweaks can be made by
that future patch.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11746