[avalanche] Improve RPC fields names and description in getavalanchepeerinfo
Summary:
The `nodes` field can be expected to be the node count but is the list of nodes, so let's call it node_list to remove confusion. Also the peerid has a different meaning than the peer id from the network log (e.g. nodeid), so let's clarify by renaming it avalanche_peerid and improve the help text a bit. There is no change in behavior.
No deprecation cycle is used because this is still an experimental feature.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, Mengerian

Reviewed By: #bitcoin_abc, Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D11758