[avalanche] Don't consider our quorum valid if we don't have enough nodes connected
Summary:
The quorum is considered valid if we got connected to enough nodes and
got some specific messages from them, but if the count drops (due to
node disconnect) we might fall below the minimum number required for
getting a convergent vote, and keep polling. This diff makes the quorum
invalid if the node has less than 8 avalanche nodes connected.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11766