Fix tsan suppression with Clang 14
Summary:
The `wallet_multiwallet --usecli` test fails consistently with TSAN and Clang 14.
It appears that the way the stack is stored prevents the suppression from triggering.
This diff uses another suppression filter which has a larger scope and faster lookup for excluding the issues from the lib.

Test Plan: Run the TSAN build with both clang 10 and 14.

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11775