[Chronik] Refactor serialization, add `db_serialize` and `db_deserialize`.
Summary:
We will use de-/serialization a lot, so having a module with nice error handling is very helpful.

The `db_` prefix indicates that this serialization is only meant for data going in and out of the database.

This also makes it much simpler to move to another serialization format if ever needed.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11798