[avalanche] Account for the avalanche outbound when checking for outbound eviction
Summary: Periodically the node checks if there is more outbound that the maximum configured and evict any extra. The count of extra outbound did not take into account the avalanche outbounds and will evict way too early. This diff fixes this issue and add a unit test for the GetExtraFullOutboundCount() method.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11803