[Cashtab] Hide spinner in Send.js Amount input on Firefox
Summary: Related to task T2581. User noted that when using firefox, the Amount input in Send.js shows the scroll spinner. Changed AntdFormCss in EnhancedInputs.js to correct this on both the Send.js and SendToken screens.

Test Plan:
In firefox:
Navigate to cashtab.com
Observe that the spinner renders in both the Send screen and the SendToken screen

In IDE:
cd web/cashtab && npm start
navigate to localhost:3000 in Firefox if FF is not your default browser
navigate to send screen - observe that the spinner does not appear
navigate to SendToken screen - observe that the spinner does not appear

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11804