[avalanche] Fix flakiness in the stale vote functional test
Summary:
The test checks for log content while waiting for the vote to stop in a loop. This can actually fail if the log message is printed while the wait funtion is timing out, causing the test to be flaky especially on slow builds. This diff fixes that by using a single wait for the proof to be dropped, while keeping the log check for ensuring it was dropped for the right reason.

Depends on D11800.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_avalanche_proof_voting

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11808