[avalanche] Don't download proofs during IBD
Summary:
We already avoid requesting compact proofs, but this extends to relayed proofs as well.
This diffs prevents the node from requesting proofs when an inv is received during IBD, and also discards the proofs received via the network in the event a peer sent us spontaneously a proof.

Depends on D11815.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11816