[avalanche] Fix flaky loops in abc_p2p_compactproofs
Summary:
There is a couple places where the test loops over mockscheduler and waits for an event. However it might occur that the mockscheduler RPC call is exectuted before the process is being rescheduled, which makes it a no op and fails the test.
Since the loops are arbitrary in this test and do not increase coverage, this diff gets rid of the potential issue by removing them. This has the added benefit of speeding up the test slightly.

Test Plan:
  for i in {1..20}; do ./test/functional/test_runner.py abc_p2p_compactproofs || break; done

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11817