[avalanche] Add coins directly instead of using coinbases in processor unit tests
Summary: Using the coinbases was efficient since they are already available in the test, but this will break if/when the dust treshold exceeds the coinbase reward. This patch replaces that implementation of GetProof() with one that adds the coins directly with the minimum amount instead of using the coinbases.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11822