Add unit test for the outbound connection logic
Summary:
This diff adds a CConnmanTest implementation that allows for testing the outbound connection logic that is used in the network thread. For now only a test for the network group limitation is added so we can later add and test avalanche specifics for this feature.

Depends on D11827.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11828