[avalanche] Fix abc_p2p_getavaaddr flakiness
Summary: The tests case that is checking for single peer request after the quorum is established can be wrong if the peers are slow enough to respond, which would cause a missing node condition during the polls and trigger a request to all our peers instead of the one. This diff fixes this by waiting for the poll to complete and run a dummy round before the actual test.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_getavaaddr

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11834