[avalanche] Don't poll blocks who are ancestors of a vote finalized block
Summary:
Once a block has been voted on and finalized, there is no point polling
its ancestors anymore.
Test is added to check the ancestors are removed from the polls, and
also that a block from a forked chain is still polled, so invalidating a
block manually is still possible if needed and won't interfere.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11849