[Cashtab] [p4 normalizeBalance fix] Change function input to nonSlpUtxos
Summary:
T2601

Depends on D11845

Part 4 in a series of diffs to improve calculation of wallet balance from utxo set

Modify the `normalizeBalance` function to accept `nonSlpUtxos` as input, instead of its parent object. The function only uses `nonSlpUtxos` so the parent object is not necessary.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D11846