[Cashtab] [p5 normalizeBalance fix] Improve name of normalizeBalance function
Summary:
T2601

Depends on D11846

Change name of `normalizeBalance` to `getWalletBalanceFromUtxos`

This is much more clear about what the function is actually doing. Not sure wtf I was thinking with the original `normalize` name.

Test Plan:
`npm test`

Note that this function is still only being referenced by the `loadStoredWallet` in `cashMethods.js`.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D11847