[Cashtab] Remove tooltip from mnemonic inputs
Summary: Related to T2606. When entering a mnemonic seed phrase, on hover the input reveals a tooltip that reads "Please include an @ in the email address..." As the mnemonic inputs are not email inputs, this should be removed. By adding a blank title to the Input tags, the tooltip will not show up

Test Plan:
cd web/cashtab && npm start
navigate to configure screen,
begin typing in the mnemonic input
hover over the input, observe that the tooltip does not appear.

open localhost in a private browser
from the onboarding page, click import wallet
begin typing in the mnemonic input
hover over the input, observe that the tooltip does not appear.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11855