[Cashtab] Hide msgs from unknown senders option
Summary: A setting should be added to allow users to hide messages from addresses not in their contacts list behind a 'show message' button. Related to T2418.

Test Plan:
cd web/cashtab && npm start
in the browser:
observe that all messages are visible by default
navigate to settings tab and toggle 'show messages' off
receive a cashtab message from an address not in contact list
receive an encrypted message from an address not in contact list
receive a message from an address that IS in contact list
receive an encrypted message from an address that IS in the contact list

from the home tab:
observe that mesages from contacts are displayed
observe that the 'show message' button is displaying for messages that aren't from contacts
click the 'show message' button on both the encrypted and unencrypted message
observe that the messages are displayed, along with a warning that the sender is not a contact

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11627