[avalanche] Poll for our local proof
Summary: So far the node was not polling for its own proof, considering it is always valid. This means that a node can start with a proof that would be rejected by the network. This diff causes the node to poll for its own proof instead so it can stop using it if it is rejected, and also it makes it appear in the finalized proofs which is more intuitive for the user.

Test Plan:
  ninja all check-extended

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D11867