[Cashtab] Patch overflow from error string in send one to many input
Summary: related to T2610. When inputting an incorrect eCash address in the DestinationAddressMulti input, a long string caused the error message to extend beyond the boundaries of the app. Container div added to avoid this issue.

Test Plan:
cd web/cashtab && npm start
navigate to send screen
in the advanced dropdown, toggle the multiple recipients switch
input a long string that is not a valid XEC address
observe that the error message does not extend beyond the boundary of the app

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11874