[Cashtab] Hide amount in error msg if it is undefined DestinationAddressMulti
Summary: Related to T2615. If the user inputs an invalid XEC address in the DestinationAddressMulti input, the error message should not include the amount param if it is undefined.

Test Plan:
cd web/cashtab && npm start
navigate to Send screen
Under the advanced tab, toggle the Multiple Recipients toggle
Input an invalid XEC address
observe that the error message does not read 'undefined' following the comma if the value is undefined

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11877