[Cashtab] Patch selectedCurrency ticker reading as XEC in send confirmation modal
Summary: Related to T2617. When sending USD, the confirmation prompt will incorrectly show XEC. This diff changes the ticker used in the modal to match with the selectedCurrency state variable.

Test Plan:
cd web/cashtab && npm start
navigate to settings screen and ensure send confirmations are turned on
navigate to Send screen
input a valid XEC address in the Send To field
in the amount drop down, select USD as the currency to send
input a valid amount in the Amount field
hit send
observe that the confirmation modal now displays the correct currency ticker

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11878