p2p: Make block stalling timeout adaptive
Summary:
This makes the stalling detection mechanism (previously a fixed
timeout of 2s) adaptive:
If we disconnect a peer for stalling, double the timeout for the
next peer - and let it slowly relax back to its default
value each time the tip advances. (Idea by Pieter Wuille)

This makes situations more unlikely in which we'd keep on
disconnecting many of our peers for stalling, even though our
own bandwidth is insufficient to download a block in 2 seconds.

Co-authored-by: Vasil Dimov <vd@FreeBSD.org>

This is a backport of [ [[https://github.com/bitcoin/bitcoin/pull/25880 | core#25880]] and [[https://github.com/bitcoin/bitcoin/pull/26982 | core#26982]]
Note that we need to specify `version=4` when calling `create_block` in the test because we miss [[https://github.com/bitcoin/bitcoin/pull/16333 | core#16333]] (which make 4 the default version for create_block)

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15080