[Cashtab] [p1 match wif to address] Align sendXec()s XEC utxo signing approach
Summary:
As per T2599, this is part of a stacked diff to sign utxos for non-eToken txs by matching wif to address.
This diff aligns the approach in the sendXec() function.

Test Plan:
npm test
send a 1 to 1 XEC tx and validate outputs in explorer
send a 1 to many XEC tx and validate outputs in explorer

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11854