[Cashtab] Replace direct svg imports with CustomIcons imports
Summary: Related to T2637. Icons should be imported through CustomIcons.js, not directly from the assets dir.

Test Plan:
cd web/cashtab && npm start
alter SignIn.js and ProtectableComponentWrapper.js to allow the FingerPrintSvg icon to appear
Ensure that it is the correct size, color, etc.

navigate to the configure screen
ensure that the trashcan icon and the edit icon are the correct size, color, and maintain the hover effect.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11932