[Cashtab] [chronik utxo handling p3] Add a function to create a preliminary `tokens` wallet state object, before getting token info
Summary:
T2447

Depends on D11933

Add a function to create a preliminary version of the `tokens` wallet state object.

Preliminary because you will need another function to get token info and calculate the true balance based on token decimals.

The way this is done is slightly different between `bch-api` and `chronik`, so splitting it up for simplicity of review.

Test Plan:
`npm test`
`npm start` and confirm the preliminary tokens object shows all of your tokens

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D11934