[avalanche] Fix abc_rpc_avalancheproof on slow machines
Summary: Since proof replacement is enabled by default the conflicting proof should have a sequence number less than the proof it replaces to trigger a conflict error. It works nontheless most of the time because of the conflicting cooldown, so disable this cooldown to make sure we are testing what we expect.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_avalancheproof

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11939