[avalanche] Fix potential wrong error message in the sendavalancheproof RPC
Summary:
The RPC does a lot of checks before attempting to register a proof and assumes it can only fail due to conflicts. This is no longer true since previously evicted dangling proofs are rejected.
This diff makes the error message more generic so the correct error message is returned, and new errors will be printed correctly as well.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11945