[Cashtab] Add hamburger menu to taskbar
Summary: Related to T2376. Added menu icon that animates from hamburger to X on click, replaced the cog in the navbar with the hamburger menu. Clicking hamburger menu icon expands the nav menu.

Test Plan:
cd web/cashtab && npm start
observe the hamburger icon in the navbar
click the hamburger icon
observe that the icon animates to an X
observe the NavMenu transition height
hover over the settings NavItem
observe that the color turns to blue on hover
click the settings NavItem
observe that the NavMenu closes on click
observe that clicking the NavItem will redirect to Settings page.

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: bytesofman, emack, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11997